Separate schema modification from gatsby-node
