+++ src/utils/create-resolvers.ts
@@ -1,0 +1,10 @@
+import { CreateResolversArgs } from "gatsby";
+
+export function createResolvers({ createResolvers }: CreateResolversArgs) {
+  createResolvers({
+    Query: {
+      site: { type: "Site!" },
+      markdownRemark: { type: "MarkdownRemark!" },
+    },
+  });
+}
